Add files via upload
We developed an ontology capable of mapping all kinds of recipes available on the Internet. To that end, we designed a comprehensive ontology that includes all the possible classes, their relations, and properties needed for recipes. In the initial step, we used the recipes dataset to define classes, sub-classes, and relationships. We also reviewed existing foods and recipes ontology to understand their composition.